Handling specific Quaternion input to avoid strange behaviour on output rotation.
Some input devices does not provide continuous quaternion values when rotating.

We know that a quaternion (x, y, z, w) represent the same rotation that (-x, -y, -z, -w).
So an input device has to send only one of the two quaternion to the software (assuming only normalized quaternions).
However, hardware like HTC Vive (for example) decided to send only quaternions with a positive y value.
This cause the w, x and z variables to jump to an opposite value when y reach 0, and the filtered quaternion to cause a strange behabiour to the filtered object (it performs a quick 360 degree rotation).

How the commited code works:
It detects if the input quaternion is in the opposite part of the space of possible values, from the current filtered quaternion.
To do so, it computes the distance between the 2 quaternions (converted to normalized Vector4).
If the distance is above sqrt(2), the input quaternion is reversed to maintain continuous input (and coherent output) for the 1 euro filter.

This commit does not affect normal behaviour (when the input use the full range of possible values)